Moved generic test classes to main namespace
